split view file
