+++ django/food/views/__init__.py
@@ -1,0 +1,4 @@
+from .views_other import *
+from .views_foodservice import *
+from .views_registration import *
+from .views_user_account import *
