+++ django/food/views/views_other.py
@@ -1,0 +1,12 @@
+from django.views.generic import TemplateView
+
+from ..forms import *
+
+class IndexView(TemplateView):
+    template_name = "food/index.html"
+
+
+    
+
+
+
